Move DiagnosticsCard, clean up localization
