+++ locales/en/public.json
@@ -138,6 +138,12 @@
       "CONTENT": "Severity scores are calculated based on the number of JFR events that were triggered by the application in the time the report was generated."
     }
   },
+  "DiagnosticsCard": {
+    "DIAGNOSTICS_CARD_DESCRIPTION": "Perform diagnostic operations on the target.",
+    "DIAGNOSTICS_CARD_DESCRIPTION_FULL": "Perform diagonstic operations from a list of supported operations on the target.",
+    "DIAGNOSTICS_CARD_TITLE": "Diagnostics",
+    "DIAGNOSTICS_GC_BUTTON": "Start Garbage Collection"
+  },
   "CHART_CARD": {
     "BUTTONS": {
       "CREATE": {
@@ -157,10 +163,6 @@
     "MBEAN_METRICS_CARD_DESCRIPTION": "Display common performance metrics from current MBean data.",
     "MBEAN_METRICS_CARD_DESCRIPTION_FULL": "Display a single performance metric from a list of supported MBeans.",
     "MBEAN_METRICS_CARD_TITLE": "MBean Metrics Chart",
-    "DIAGNOSTICS_CARD_DESCRIPTION": "Perform diagnostic operations on the target.",
-    "DIAGNOSTICS_CARD_DESCRIPTION_FULL": "Perform diagonstic operations from a list of supported operations on the target.",
-    "DIAGNOSTICS_CARD_TITLE": "Diagnostics",
-    "DIAGNOSTICS_GC_BUTTON": "Start Garbage Collection",
     "NO_RECORDING": {
       "DESCRIPTION": "Metrics cards display data taken from running Flight Recordings with the label <label>origin={{recordingName}}</label>. No such Recordings are currently available.",
       "TITLE": "No source Recording"
