+++ src/app/Dashboard/utils.tsx
@@ -27,7 +27,7 @@ import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDispatch } from 'react-redux';
 import { AutomatedAnalysisCardDescriptor } from './AutomatedAnalysis/AutomatedAnalysisCard';
-import { DiagnosticsCardDescriptor } from './Charts/diagnostics/DiagnosticsCard';
+import { DiagnosticsCardDescriptor } from './Diagnostics/DiagnosticsCard';
 import { JFRMetricsChartCardDescriptor } from './Charts/jfr/JFRMetricsChartCard';
 import { MBeanMetricsChartCardDescriptor } from './Charts/mbean/MBeanMetricsChartCard';
 import { JvmDetailsCardDescriptor } from './JvmDetails/JvmDetailsCard';
